+++ detect-feature/src/lib.rs
@@ -1,5 +1,0 @@
-#![feature(stdarch_mips_feature_detection)]
-#![feature(stdarch_arm_feature_detection)]
-#![feature(stdarch_powerpc_feature_detection)]
-
-include!(concat!(env!("OUT_DIR"), "/lib.rs"));